Fix issue converting Unwrap return value to bool improperly
Unwrap returns negative error code on failure, or data size on success. Thus, to convert it to a "success" bool, it should be tested for negativity.

Returning Unwrap() as a bool will report success for errors and failure for success with zero data size.

To correctly convert Unwrap() to a boolean success value, (Unwrap() >= 0) should be used.